+++ src/XrmContext/CommandLine/Arguments.fs
@@ -69,15 +69,20 @@ type Args private () =
       description="Url to the Organization.svc"
       required=true }
 
+    { command="method"
+      altCommands=[]
+      description="Connection method"
+      required=false }
+
     { command="username"
       altCommands=["u"; "usr"]
       description="CRM Username"
-      required=true }
+      required=false }
 
     { command="password"
       altCommands=["p"; "pwd"]
       description="CRM Password"
-      required=true }
+      required=false }
 
     { command="domain"
       altCommands=["d"; "dmn"]
@@ -97,6 +102,11 @@ type Args private () =
     { command="mfaReturnUrl"
       altCommands=[]
       description="Return URL defined for the Azure Application"
+      required=false }
+
+    { command="mfaClientSecret"
+      altCommands=[]
+      description="Client secret for the Azure Application"
       required=false }
   ]