+++ RELEASE_NOTES.md
@@ -1,4 +1,13 @@
 # Release Notes
+### 1.7.2 - February 25 2020
+* Fixed error when an alternate key used an optionset attribute
+
+### 1.7.1 - February 13 2020
+* Fix assembly reference issue when calling from Daxif
+
+### 1.7.0 - February 11 2020
+* Added support for client secret authentication
+
 ### 1.6.0 - September 24 2019
 * Added support for MFA